Install SKD
It is mentioned on the [SDK](https://github.com/elk-audio/elkpi-sdk)
repo that this container could be used together with the SDK
for example to cross-compile plugins.

It would make sense to install the SDK already at build time of
the container to avoid having to install it every time when the
container is started.

As an alternative, this could also be added in another `Dockerfile`
which uses the elk-audio-os-builder as a base container.